+++ package.json
@@ -92,12 +92,10 @@
     "@angular/platform-browser": "^14.2.2",
     "@angular/platform-browser-dynamic": "^14.2.2",
     "@angular/router": "^14.2.2",
-    "@types/validator": "^13.7.7",
     "husky": "^8.0.1",
     "open-props": "^1.4.16",
     "rxjs": "~7.5.6",
     "tslib": "^2.4.0",
-    "validator": "^13.7.0",
     "vest": "^4.6.7",
     "wireit": "^0.7.1",
     "zone.js": "~0.11.8"
